Add GLICOL, kilobeat and Web Audio Conference
GLICOL is a brand new livecoding language written in Rust which has a browser-based IDE and runs on the AudioWorklet engine. kilobeat is a livecoding collaborative instrument inspired by bytebeat with some cool insteresting and recording features. The Web Audio Conference is addressing some topics related to live and creative coding (and where I first heard of GLICOL and kilobeat).